+++ src/main/java/ro/fiipractic/mycinema/dtos/ReservationDto.java
@@ -6,9 +6,9 @@ public class ReservationDto {
 
     private Integer numberOfTickets;
 
-    private Integer customer_id;
+    private Long customer_id;
 
-    private Integer movie_instance_id;
+    private Long movie_instance_id;
 
     public Long getId() {
         return id;
@@ -26,19 +26,19 @@ public class ReservationDto {
         this.numberOfTickets = numberOfTickets;
     }
 
-    public Integer getCustomer_id() {
+    public Long getCustomer_id() {
         return customer_id;
     }
 
-    public void setCustomer_id(Integer customer_id) {
+    public void setCustomer_id(Long customer_id) {
         this.customer_id = customer_id;
     }
 
-    public Integer getMovie_instance_id() {
+    public Long getMovie_instance_id() {
         return movie_instance_id;
     }
 
-    public void setMovie_instance_id(Integer movie_instance_id) {
+    public void setMovie_instance_id(Long movie_instance_id) {
         this.movie_instance_id = movie_instance_id;
     }
 }
